+++ CHANGELOG.md
@@ -9,6 +9,13 @@ See [GitHub releases](https://github.com/mll-lab/php-utils/releases).
 
 ## Unreleased
 
+## v2.0.0
+
+### Changed
+
+- Generate the Reads-Sections dynamically from the OverrideCycles-part of the Samples-Section for Illumina NovaSeq Sample Sheets (V2)
+- OverrideCycles on BclSample not nullable
+
 ## v1.14.0
 
 ### Added